refactor: extract dependency
