--- examples/synthemol/features/__init__.py
@@ -1,63 +1,0 @@
-from .features_generators import get_available_features_generators
-from .features_generators import get_features_generator
-from .features_generators import morgan_binary_features_generator
-from .features_generators import morgan_counts_features_generator
-from .features_generators import rdkit_2d_features_generator
-from .features_generators import rdkit_2d_normalized_features_generator
-from .features_generators import register_features_generator
-from .featurization import BatchMolGraph
-from .featurization import MolGraph
-from .featurization import atom_features
-from .featurization import atom_features_zeros
-from .featurization import bond_features
-from .featurization import get_atom_fdim
-from .featurization import get_bond_fdim
-from .featurization import is_adding_hs
-from .featurization import is_explicit_h
-from .featurization import is_mol
-from .featurization import is_reaction
-from .featurization import map_reac_to_prod
-from .featurization import mol2graph
-from .featurization import onek_encoding_unk
-from .featurization import reset_featurization_parameters
-from .featurization import set_adding_hs
-from .featurization import set_explicit_h
-from .featurization import set_extra_atom_fdim
-from .featurization import set_extra_bond_fdim
-from .featurization import set_reaction
-from .utils import load_features
-from .utils import load_valid_atom_or_bond_features
-from .utils import save_features
-
-__all__ = [
-    "get_available_features_generators",
-    "get_features_generator",
-    "morgan_binary_features_generator",
-    "morgan_counts_features_generator",
-    "rdkit_2d_features_generator",
-    "rdkit_2d_normalized_features_generator",
-    "atom_features",
-    "atom_features_zeros",
-    "bond_features",
-    "BatchMolGraph",
-    "get_atom_fdim",
-    "set_extra_atom_fdim",
-    "get_bond_fdim",
-    "set_extra_bond_fdim",
-    "set_explicit_h",
-    "set_adding_hs",
-    "set_reaction",
-    "is_reaction",
-    "is_explicit_h",
-    "is_adding_hs",
-    "is_mol",
-    "mol2graph",
-    "map_reac_to_prod",
-    "MolGraph",
-    "onek_encoding_unk",
-    "load_features",
-    "save_features",
-    "load_valid_atom_or_bond_features",
-    "reset_featurization_parameters",
-    "register_features_generator",
-]
